fix: add transaction

+++ back-end/client/raja-ongkir/model.go
@@ -22,17 +22,17 @@ type Destination struct {
 }
 
 type ShippingCost struct {
-	ShippingName     string `json:"shipping_name"`
-	ServiceName      string `json:"service_name"`
-	Weight           int    `json:"weight"`
-	IsCOD            bool   `json:"is_cod"`
-	ShippingCost     int    `json:"shipping_cost"`
-	ShippingCashback int    `json:"shipping_cashback"`
-	ShippingCostNet  int    `json:"shipping_cost_net"`
-	GrandTotal       int    `json:"grandtotal"`
-	ServiceFee       int    `json:"service_fee"`
-	NetIncome        int    `json:"net_income"`
-	ETD              string `json:"etd"`
+	ShippingName     string  `json:"shipping_name"`
+	ServiceName      string  `json:"service_name"`
+	Weight           float64 `json:"weight"`
+	IsCOD            bool    `json:"is_cod"`
+	ShippingCost     int     `json:"shipping_cost"`
+	ShippingCashback int     `json:"shipping_cashback"`
+	ShippingCostNet  int     `json:"shipping_cost_net"`
+	GrandTotal       int     `json:"grandtotal"`
+	ServiceFee       int     `json:"service_fee"`
+	NetIncome        int     `json:"net_income"`
+	ETD              string  `json:"etd"`
 }
 
 type ShippingCostResponse struct {
